requirements.txt
pandas
numpy
requests
python-telegram-bot==20.7
ta
scipy
pyyaml
